% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter-text.R
\name{tweet_autolink}
\alias{tweet_autolink}
\title{Perform tweet autolinking}
\usage{
tweet_autolink(x)
}
\arguments{
\item{x}{A character vector of tweets to be autolinked}
}
\value{
A character vector giving the HTML which would be used by Twitter
if \code{x} were posted as a tweet.
}
\description{
Perform tweet autolinking
}
\examples{
tweet_autolink(c(
  "Tweet with no links",
  "Tweet with a normal link: https://www.r-project.org",
  "Tweet with a mention @hadleywickham",
  "Tweet with a hastag: #RStats",
  "Tweet with a cashtag: $RSTATS"
))
}
