% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter-text.R
\name{tweet_info}
\alias{tweet_info}
\alias{tweet_permillage}
\alias{tweet_length}
\alias{tweet_is_valid}
\alias{tweet_get_mentions}
\alias{tweet_autolink}
\title{Wrappers from the twitter-text library}
\usage{
tweet_info(x)

tweet_permillage(x)

tweet_length(x)

tweet_is_valid(x)

tweet_get_mentions(x)

tweet_autolink(x)
}
\arguments{
\item{x}{A character vector of tweets}
}
\description{
Wrappers from the twitter-text library
}
\examples{
tweets <- c(
  "This is a first tweet. Simple!",
  "This tweet tags @hadleywickham and @_wurli",
  "This tweet links {rtweet}: https://docs.ropensci.org/rtweet/",
  strrep("This tweet is way too long! ", 20)
)

tweet_info(tweets)
tweet_permillage(tweets)
tweet_length(tweets)
tweet_is_valid(tweets)
tweet_get_mentions(tweets)
tweet_autolink(tweets)
}
